"feat: RegionRepo changes."

+++ src/main/java/gr/aueb/cf/eduapp/repository/RegionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RegionRepository extends JpaRepository<Region, Integer> {
+public interface RegionRepository extends JpaRepository<Region, Long> {
 
     List<Region> findAllByOrderByNameAsc();
 }
